requirements.txt
Django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf-yasg==1.21.7
joblib==1.3.2
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
matplotlib==3.8.2
seaborn==0.13.0
